Create Repository Interfaces. Change ID from int to Long

+++ src/test/java/com/radello/constructioncompanyorganizer/converter/BudgetCommandToBudgetTest.java
@@ -9,7 +9,8 @@ import static org.junit.jupiter.api.Assertions.*;
 
 class BudgetCommandToBudgetTest {
 
-    private static final int ID_VALUE = 1;
+
+    private static final Long ID_VALUE = 1L;
     private static final int AMOUNT_VALUE = 12;
 
     BudgetCommandToBudget budgetCommandToBudget;
